pyproject.toml
[project]
name = "pc-power-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
]
